Added music, SFX, start and end screens, multiple levels
I have added music for the background so that it is not silent and sound effects for the collection of item. another thing which I have added is the start and end screen for the game. Also. the final new addition is that we have multiple levels which can be played.